updated database, added inventory file, renamed data to warehouses
