+++ back-end/Database/inventory.json
@@ -1,0 +1,132 @@
+[
+  {
+    "id": 61132,
+    "name": "Handcrafted Plastic Tuna",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "DDXUZMI1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 48586,
+    "name": "Refined Steel Fish",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "KEMUWFS1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 99119,
+    "name": "Incredible Cotton Shoes",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "ZTNEPWG1IIV",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 59209,
+    "name": "Small Wooden Car",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "CDMEMKK1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 34804,
+    "name": "Intelligent Fresh Pizza",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "CHYULTO1180",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 49625,
+    "name": "Gorgeous Plastic Bacon",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "WMZAHNR1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 21210,
+    "name": "Licensed Plastic Pants",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "IKAUDKB1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 94384,
+    "name": "Handmade Fresh Bacon",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "KIYAIRW1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 63788,
+    "name": "Licensed Concrete Shoes",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "ECLIAMZ1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  },
+  {
+    "id": 93580,
+    "name": "Small Frozen Pants",
+    "short-description": "Here is a very brief description of the product",
+    "long-description": "Here is a more detailed summary of the product name, it's uses, inductries and possible attributes that could be used to descript the product.",
+    "ordered-by": "Mark Saunders",
+    "last-ordered": "05/24/2018",
+    "location": "Toronto, CAN",
+    "quantity": 12000,
+    "reference-number": "ECPARUA1",
+    "in-stock": true,
+    "categories": ["Industrial", "Automotive", "Heavy", "Mechanical", "Engineering", "Sales"]
+  }
+]
