+++ back-end/Routes/inventory.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const fs = require('fs')
-const data = require('../Database/data')
+const data = require('../Database/inventory')
 
 router.get('/')
